+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project changes will be shown here.
 - Issue 1308 - Added Allocate more time modal.
 - Issue 1329 - Add sales order to RequestDashboard
 - Issue 1328 - Starting SR from a call back can cause SLA to fail
+- Issue 1323 - Include Mailbox Additional Permission in Office 365 Mailbox Export
 ### Changed
 - Issue 1276 - Monitored SR email, Change Status: Awaiting Customer to Status: On Hold
 - Issue 1310 - Show sales orders on overtime page.